Moved the dbapi into commonlib
