+++ doc/sphinx/openquake.server.rst
@@ -12,14 +12,6 @@ Subpackages
 Submodules
 ----------
 
-openquake.server.dbapi module
------------------------------
-
-.. automodule:: openquake.server.dbapi
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
 openquake.server.dbserver module
 --------------------------------
